% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmsim.R
\name{convert_qmsim_to_mix99}
\alias{convert_qmsim_to_mix99}
\title{Convert QMSim Output to MiX99 Input}
\usage{
convert_qmsim_to_mix99(
  ptbl_qm = NULL,
  ps_qm_path = NULL,
  ps_out_dir = ".",
  ps_file_stem = paste0(format(Sys.time(), "\%Y\%m\%d\%H\%M\%S"), "_mix99", collapse =
    ""),
  pn_mvc = -99999,
  pl_vcmp
)
}
\arguments{
\item{ptbl_qm}{tibble with QMSim output}

\item{ps_qm_path}{QMSim output file}

\item{ps_out_dir}{Output directory for Mix99 files}

\item{ps_file_stem}{File stem for Mix99 files}

\item{pn_mvc}{numeric missing value code}
}
\description{
Convert QMSim Output to MiX99 Input
}
\examples{
s_qm_path <- system.file("extdata", "qmsim", "p1_data_001.txt", package = "qzwsmix99rt")
s_tmp_dir <- tempdir()
convert_qmsim_to_mix99(ps_qm_path = s_qm_path, ps_out_dir = file.path(s_tmp_dir, 'mix99_work'))
fs::dir_delete(s_tmp_dir)
}
