% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_example.R
\name{prepare_example_p1}
\alias{prepare_example_p1}
\title{Prepare Example Evaluation with p1 Data}
\usage{
prepare_example_p1(
  ps_work_dir = file.path(tempdir(), "work_p1"),
  ps_source_dir = system.file("extdata", "mix99", package = "qzwsmix99rt"),
  pb_force = FALSE,
  pb_debug = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_work_dir}{working directory for example files}

\item{ps_source_dir}{source directory from where example files are copied}

\item{pb_force}{forcing copy by first deleting any existing working directory}

\item{pb_debug}{run in debugging mode and producing log messages}

\item{plogger}{logr logger object}
}
\description{
The example data which are part of this package are copied into a working
directory. The example files can be used to run genetic evaluation
}
