% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qzwsmix99rt_logger.R
\name{qzwsmix99rt_log_debug}
\alias{qzwsmix99rt_log_debug}
\title{Wrapper for log4r debug}
\usage{
qzwsmix99rt_log_debug(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_caller}{function from which we are called}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r debug
}
\examples{
\dontrun{
qzwsmix99rt_logger <- get_qzwsmix99rt_logger()
qzwsmix99rt_log_level(qzwsmix99rt_logger, 'DEBUG')
qzwsmix99rt_log_debug(qzwsmix99rt_logger, 'Examples', 'test message')
}
}
