% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qzwsmix99rt_logger.R
\name{qzwsmix99rt_log_level}
\alias{qzwsmix99rt_log_level}
\title{Wrapper to set the level of a logger}
\usage{
qzwsmix99rt_log_level(
  plogger,
  ps_level = c("DEBUG", "INFO", "WARN", "ERROR", "FATAL")
)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_level}{new level of plogger}
}
\description{
Wrapper to set the level of a logger
}
\examples{
\dontrun{
qzwsmix99rt_logger <- get_qzwsmix99rt_logger()
qzwsmix99rt_log_level(qzwsmix99rt_logger, 'INFO')
}
}
