% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmsim.R
\name{read_qmsim}
\alias{read_qmsim}
\title{Read Output File of QMSim}
\usage{
read_qmsim(ps_path)
}
\arguments{
\item{ps_path}{path to QMSim output file}
}
\value{
tibble with QMSim output
}
\description{
Read Output File of QMSim
}
\examples{
s_qmsim_path <- system.file("extdata", "qmsim", "p1_data_001.txt", package = "qzwsmix99rt")
tbl_qmsim <- read_qmsim(ps_path = s_qmsim_path)
}
