% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mix99.R
\name{run_mix99}
\alias{run_mix99}
\title{Run MiX99 Evaluation}
\usage{
run_mix99(
  ps_work_dir,
  ps_log_file = file.path(ps_work_dir, paste0(format(Sys.time(), "\%Y\%m\%d\%H\%M\%S"),
    "_mix99.log", collapse = ""))
)
}
\arguments{
\item{ps_work_dir}{working directory containing data, pedigree and clim file}

\item{ps_log_file}{log file to which mix99 output is written}
}
\description{
Given a working directory with data file, pedigree file and clim file, the
mix99 program is executed using the system() function. The output is captured
and written to a log file
}
\examples{
\dontrun{
test_dir <- prepare_example_p1()
run_mix99(ps_work_dir = test_dir)
fs::dir_delete(test_dir)
}
}
